增加payApply接口，此接口还需要仔细测试

+++ jpress/jpress-model/src/main/java/yjt/model/base/BaseContract.java
@@ -3,7 +3,6 @@
  */
 package yjt.model.base;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 
 import com.jfinal.plugin.activerecord.IBean;
@@ -138,7 +137,7 @@ public class BaseContract<M extends BaseContract<M>> extends JModel<M> implement
 		set("contract_type_id", contractTypeId);
 	}
 
-	public Long getContractTypeId() {
+	public int getContractTypeId() {
 		return get("contract_type_id");
 	}
 
@@ -146,28 +145,28 @@ public class BaseContract<M extends BaseContract<M>> extends JModel<M> implement
 		set("credit_id", creditId);
 	}
 
-	public Long getCreditId() {
+	public BigInteger getCreditId() {
 		return get("credit_id");
 	}
 
 	public void setDebitId(java.math.BigInteger debitId){
 		set("debit_id", debitId);
 	}
-	public Long getDebitId(){
+	public BigInteger getDebitId(){
 		return get("debit_id");
 	}
 	
 	public void setAmount(double amount){
 		set("amount", amount);
 	}
-	public BigDecimal getAmount(){
+	public Double getAmount(){
 		return get("amount");
 	}
 	
 	public void setAnnualRate(double annualRate ){
 		set("annual_rate", annualRate);
 	}
-	public BigDecimal getAnnualRate(){
+	public Double getAnnualRate(){
 		return get("annual_rate");
 	}
 	
@@ -228,7 +227,7 @@ public class BaseContract<M extends BaseContract<M>> extends JModel<M> implement
 	public void setAppovalU1(int approvalU1){
 		set("approval_u1", approvalU1);
 	}
-	public Long getAppovalU1(){
+	public BigInteger getAppovalU1(){
 		return get("approval_u1");
 	}
 	public void setApprovalTime1(java.util.Date approvalTime1 ){
@@ -241,7 +240,7 @@ public class BaseContract<M extends BaseContract<M>> extends JModel<M> implement
 	public void setAppovalU2(int approvalU2){
 		set("approval_u2", approvalU2);
 	}
-	public Long getAppovalU2(){
+	public BigInteger getAppovalU2(){
 		return get("approval_u2");
 	}
 	public void setApprovalTime2(java.util.Date approvalTime2 ){
@@ -254,7 +253,7 @@ public class BaseContract<M extends BaseContract<M>> extends JModel<M> implement
 	public void setAppovalU3(int approvalU3){
 		set("approval_u3", approvalU3);
 	}
-	public Long getAppovalU3(){
+	public BigInteger getAppovalU3(){
 		return get("approval_u3");
 	}
 	public void setApprovalTime3(java.util.Date approvalTime3 ){
